refactor: split relational models and schemas by domain
